pyproject.toml
[project]
name = "ragenius-chatbot"
version = "0.1.0"
description = "this project for analyses documents like pdf and images"
authors = [
    {name = "issam adoch"}
]
packages = [{ include = "app" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "langchain[community,ollama] (>=1.2.13,<2.0.0)",
    "langgraph[postgres] (>=1.1.3,<2.0.0)",
    "langgraph-checkpoint-postgres (>=2.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "pypdf (>=6.9.2,<7.0.0)",
    "opencv-python-headless (>=4.13.0.92,<5.0.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "langchain-mistralai (>=1.1.2,<2.0.0)",
    "google-auth (>=2.49.1,<3.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
